refactor: move module augmentation to a dedicated d.ts file
